+++ src/main/java/com/ndb/auction/models/transactions/paypal/PaypalAuctionTransaction.java
@@ -14,7 +14,7 @@ public class PaypalAuctionTransaction extends PaypalDepositTransaction {
     public PaypalAuctionTransaction (
         int userId, 
         int auctionId, 
-        Long amount,
+        Double amount,
         String paypalOrderid,
         String payaplOrderStatus
     ) {
