+++ src/main/java/com/ndb/auction/resolver/payment/presale/PresaleStripe.java
@@ -20,7 +20,7 @@ import graphql.kickstart.tools.GraphQLQueryResolver;
 public class PresaleStripe extends BaseResolver implements GraphQLQueryResolver, GraphQLMutationResolver {
 
     @PreAuthorize("isAuthenticated()")
-    public PayResponse payStripeForPreSale(int presaleId, int orderId, Long amount, String paymentIntentId, String paymentMethodId, boolean isSaveCard) {
+    public PayResponse payStripeForPreSale(int presaleId, int orderId, Double amount, String paymentIntentId, String paymentMethodId, boolean isSaveCard) {
         UserDetailsImpl userDetails = (UserDetailsImpl) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
         int userId = userDetails.getId();
         StripePresaleTransaction m = new StripePresaleTransaction(userId, presaleId, orderId, amount, paymentIntentId, paymentMethodId);
@@ -28,7 +28,7 @@ public class PresaleStripe extends BaseResolver implements GraphQLQueryResolver,
     }
 
     @PreAuthorize("isAuthenticated()")
-    public PayResponse payStripeForPreSaleWithSavedCard(int presaleId, int orderId, Long amount, int cardId, String paymentIntentId) {
+    public PayResponse payStripeForPreSaleWithSavedCard(int presaleId, int orderId, Double amount, int cardId, String paymentIntentId) {
         UserDetailsImpl userDetails = (UserDetailsImpl) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
         int userId = userDetails.getId();
         StripeCustomer customer = stripeCustomerService.getSavedCard(cardId);
